Создание презентера точек. Начло переноса кода.
